requirements.txt
Django==4.2.4
ipython==8.14.0
Pillow==10.0.0
psycopg2-binary==2.9.7
pytz==2023.3
rest_framework_simplejwt